reformatting, and break up of placeIdentification

+++ paice.cpp
@@ -6,7 +6,7 @@
  - introduced trailing 2-digit strings as state markers in rules,
  - added rule display and debugging option,
  - modified applyrule() by prescreening rules to return "not applicable"
-   for matches that would create unacceptably short stems,
+	 for matches that would create unacceptably short stems,
  - added return code to readrules().
 */
 /* 8/10/2003 - AZ - modified stem() to return the rule trace */
@@ -29,218 +29,224 @@
 #include <sstream>
 
 #define maxlinelength 1024 /* Maximum length of line read from file */
-enum states { s_notapply,s_stop,s_continue };
+enum states { s_notapply, s_stop, s_continue };
 vector <cStemmer::cSuffixRule> cStemmer::stemRules;
 vector <cStemmer::tPrefixRule> cStemmer::prefixRules;
 unordered_set<int> cStemmer::unacceptableCombinationForms;
 
 /* * * APPLYRULE()  * * * * * * * */
-int cStemmer::applyStemRule(wstring word,cSuffixRule rule,vector <cSuffixRule> &rulesUsed,cIntArray trail)
-{ LFS
-  /* Apply the rule r to word,leaving results in word.Return stop,continue */
-  /* or notapply as appropriate */
+int cStemmer::applyStemRule(wstring word, cSuffixRule rule, vector <cSuffixRule>& rulesUsed, cIntArray trail)
+{
+	LFS
+		/* Apply the rule r to word,leaving results in word.Return stop,continue */
+		/* or notapply as appropriate */
 
-  if(rulesUsed.size() && rule.intact) /* If it should be the first rule applied but isn't... */
-    return s_notapply; /* ..then rule fails */
+		if (rulesUsed.size() && rule.intact) /* If it should be the first rule applied but isn't... */
+			return s_notapply; /* ..then rule fails */
 
-  int stemlen = word.length() - rule.keystr.length(); /* Find where suffix  should start */
+	int stemlen = word.length() - rule.keystr.length(); /* Find where suffix  should start */
 
-  /* AZ - the following change speeds up the process:
-  1) Avoid matching rules for which the suffix is longer than the word.
-  2) Avoid applying rules that would result in unacceptably short stems.
-    The second condition  makes it possible to apply rules of
-    smaller length that are substrings of another rule.
-    E.g, given MINSTEMSIZE = 3 and the rules:
-      tions,?,stop
-      ions,?,stop
-      s,?,stop
-    For the word "actions"  'tions' would have matched and generated
-      the 2 letter stem "ac".  However, since this is smaller than
-      MINSTEMSIZE, the rule is considered not applicable and the next
-      rule, 'ions' will apply and generate "act")
-    For "lions", 'ions' won't apply, but 's' applies to generate 'lion'
-  */
-  if ((stemlen < 0) || (stemlen + rule.repstr.length() < MINSTEMSIZE) || rule.keystr!=word.substr(stemlen,word.length()-stemlen)) /* AZ change */
-    return s_notapply;
+	/* AZ - the following change speeds up the process:
+	1) Avoid matching rules for which the suffix is longer than the word.
+	2) Avoid applying rules that would result in unacceptably short stems.
+		The second condition  makes it possible to apply rules of
+		smaller length that are substrings of another rule.
+		E.g, given MINSTEMSIZE = 3 and the rules:
+			tions,?,stop
+			ions,?,stop
+			s,?,stop
+		For the word "actions"  'tions' would have matched and generated
+			the 2 letter stem "ac".  However, since this is smaller than
+			MINSTEMSIZE, the rule is considered not applicable and the next
+			rule, 'ions' will apply and generate "act")
+		For "lions", 'ions' won't apply, but 's' applies to generate 'lion'
+	*/
+	if ((stemlen < 0) || (stemlen + rule.repstr.length() < MINSTEMSIZE) || rule.keystr != word.substr(stemlen, word.length() - stemlen)) /* AZ change */
+		return s_notapply;
 
-  /* If ending matches key string.. */
-  if (rule.protect) return s_stop; /* If it is protected,then stop */
+	/* If ending matches key string.. */
+	if (rule.protect) return s_stop; /* If it is protected,then stop */
 
-  /* Before replacing, exclude terminal state markers from the
-    replacement string length count to see if stem will be valid.
-    It is more efficient to check here than before the match. */
-  int rl=rule.repstr.length();
-  if (rl >= 2 && iswdigit(rule.repstr[rl-1]) && iswdigit(rule.repstr[rl-2]))  rl-=2;
-  if (stemlen + rl < MINSTEMSIZE ) return s_notapply;
+	/* Before replacing, exclude terminal state markers from the
+		replacement string length count to see if stem will be valid.
+		It is more efficient to check here than before the match. */
+	int rl = rule.repstr.length();
+	if (rl >= 2 && iswdigit(rule.repstr[rl - 1]) && iswdigit(rule.repstr[rl - 2]))  rl -= 2;
+	if (stemlen + rl < MINSTEMSIZE) return s_notapply;
 
-  /* Replace matching keystr with repstr. */
-  rule.text=word.substr(0,stemlen)+rule.repstr;
-  rule.trail=trail;
-  rulesUsed.push_back(rule);
-  #ifdef LOG_DICTIONARY
-  lplog(L"rule #%d applied to %s resulting in %s trail=%s.",rule.rulenum,word.c_str(),rule.text.c_str(),trail.concatToString().c_str());
-  #endif
-  return (rule.cont) ? s_continue:s_stop;/* If continue flag is set,return cont */
+	/* Replace matching keystr with repstr. */
+	rule.text = word.substr(0, stemlen) + rule.repstr;
+	rule.trail = trail;
+	rulesUsed.push_back(rule);
+#ifdef LOG_DICTIONARY
+	lplog(L"rule #%d applied to %s resulting in %s trail=%s.", rule.rulenum, word.c_str(), rule.text.c_str(), trail.concatToString().c_str());
+#endif
+	return (rule.cont) ? s_continue : s_stop;/* If continue flag is set,return cont */
 }
 
-int cStemmer::getInflectionNum(wchar_t const *inflection)
-{ LFS
-  if (!inflection[0]) return 0;
-  int temp=0;
-  const wchar_t *ch;
-  tInflectionMap *inflectionMaps[]={nounInflectionMap,verbInflectionMap,adjectiveInflectionMap,adverbInflectionMap};
-  for (int map=0; map<4; map++)
-    for (int I=0; inflectionMaps[map][I].num>=0; I++)
-    {
-      const wchar_t *name=inflectionMaps[map][I].name;
-      if ((ch=wcsstr(inflection,name)) && (*(ch+wcslen(name))==L' ' || !*(ch+wcslen(name))))
-        temp|=inflectionMaps[map][I].num;
-    }
-  return temp;
+int cStemmer::getInflectionNum(wchar_t const* inflection)
+{
+	LFS
+		if (!inflection[0]) return 0;
+	int temp = 0;
+	const wchar_t* ch;
+	tInflectionMap* inflectionMaps[] = { nounInflectionMap,verbInflectionMap,adjectiveInflectionMap,adverbInflectionMap };
+	for (int map = 0; map < 4; map++)
+		for (int I = 0; inflectionMaps[map][I].num >= 0; I++)
+		{
+			const wchar_t* name = inflectionMaps[map][I].name;
+			if ((ch = wcsstr(inflection, name)) && (*(ch + wcslen(name)) == L' ' || !*(ch + wcslen(name))))
+				temp |= inflectionMaps[map][I].num;
+		}
+	return temp;
 }
 
 int cStemmer::readStemRules(void)
-{ LFS
-  /* Format is: keystr,repstr,flags where keystr and repstr are strings,and */
-  /* flags are:"protect","intact","continue" (without the inverted commas in the actual file).  */
+{
+	LFS
+		/* Format is: keystr,repstr,flags where keystr and repstr are strings,and */
+		/* flags are:"protect","intact","continue" (without the inverted commas in the actual file).  */
 
-  FILE *fp=_wfopen(L"source\\lists\\suffixRules.txt",L"rb");
-  if (!fp) 
+		FILE* fp = _wfopen(L"source\\lists\\suffixRules.txt", L"rb");
+	if (!fp)
 	{
-		lplog(LOG_FATAL_ERROR,L"Suffix file not found.");
+		lplog(LOG_FATAL_ERROR, L"Suffix file not found.");
 		return NO_SUFFIX_RULES_FILE;
 	}
-  wchar_t s[maxlinelength];
-  cSuffixRule temp;
-  int line;
-  /* Read a line at a time until eof */
-  for (line=1; fgetws(s,maxlinelength,fp); line++)
-  {
-		if (s[0]==0xFEFF) // detect BOM
-			memcpy(s,s+1,wcslen(s+1));
-    if ((s[0] == L';') || (s[0] == L'\r') || (s[0] == L'\n') || (s[0] == L' '))
-      continue;
-	  wchar_t savecopy[maxlinelength];
-		wcscpy(savecopy,s);
-    wchar_t *ch=wcschr(s,L','),*savech;
-    if (!ch) 
+	wchar_t s[maxlinelength];
+	cSuffixRule temp;
+	int line;
+	/* Read a line at a time until eof */
+	for (line = 1; fgetws(s, maxlinelength, fp); line++)
+	{
+		if (s[0] == 0xFEFF) // detect BOM
+			memcpy(s, s + 1, wcslen(s + 1));
+		if ((s[0] == L';') || (s[0] == L'\r') || (s[0] == L'\n') || (s[0] == L' '))
+			continue;
+		wchar_t savecopy[maxlinelength];
+		wcscpy(savecopy, s);
+		wchar_t* ch = wcschr(s, L','), * savech;
+		if (!ch)
 		{
-		  fclose(fp);
-			lplog(LOG_FATAL_ERROR,L"Error parsing (0) suffix rule on line %d: %s",line,savecopy);
+			fclose(fp);
+			lplog(LOG_FATAL_ERROR, L"Error parsing (0) suffix rule on line %d: %s", line, savecopy);
 			return SUFFIX_RULES_PARSE_ERROR;
 		}
-    *ch=0;
-    temp.keystr = s; /* Copy key string into the rule struct */
-    if (!(ch=wcschr(savech=ch+1,L','))) 
+		*ch = 0;
+		temp.keystr = s; /* Copy key string into the rule struct */
+		if (!(ch = wcschr(savech = ch + 1, L',')))
 		{
-		  fclose(fp);
-			lplog(LOG_FATAL_ERROR,L"Error parsing (1) suffix rule on line %d: %s",line,savecopy);
+			fclose(fp);
+			lplog(LOG_FATAL_ERROR, L"Error parsing (1) suffix rule on line %d: %s", line, savecopy);
 			return SUFFIX_RULES_PARSE_ERROR;
 		}
-    *ch=0;
-    temp.repstr = savech;
-    if (!(ch=wcschr(savech=ch+1,L','))) 
+		*ch = 0;
+		temp.repstr = savech;
+		if (!(ch = wcschr(savech = ch + 1, L',')))
 		{
-		  fclose(fp);
-			lplog(LOG_FATAL_ERROR,L"Error parsing (2) suffix rule on line %d: %s",line,savecopy);
+			fclose(fp);
+			lplog(LOG_FATAL_ERROR, L"Error parsing (2) suffix rule on line %d: %s", line, savecopy);
 			return SUFFIX_RULES_PARSE_ERROR;
 		}
-    *ch=0;
-    temp.form = savech;
-    if (!(ch=wcschr(savech=ch+1,L','))) 
+		*ch = 0;
+		temp.form = savech;
+		if (!(ch = wcschr(savech = ch + 1, L',')))
 		{
-		  fclose(fp);
-			lplog(LOG_FATAL_ERROR,L"Error parsing (3) suffix rule on line %d: %s",line,savecopy);
+			fclose(fp);
+			lplog(LOG_FATAL_ERROR, L"Error parsing (3) suffix rule on line %d: %s", line, savecopy);
 			return SUFFIX_RULES_PARSE_ERROR;
 		}
-    *ch=0;
-    temp.inflection = getInflectionNum(savech);
-    if ((ch=wcschr(savech=ch+1,L';'))) *ch=0;
-    temp.protect = wcsstr(savech,L"protect")!=NULL;
-    temp.intact = wcsstr(savech,L"intact")!=NULL;
-    temp.cont = wcsstr(savech,L"continue")!=NULL;
-    temp.rulenum = line; /* Line number of rule in file */
-    /* Check replacement string for special 2-digit markers  */
-    int rl = temp.repstr.length();
-    if (rl > 1 && iswdigit(temp.repstr[rl-1]) && iswdigit(temp.repstr[rl-2]) && temp.cont == false)
-      lplog(L"** WARNING ** ReadRules: State marker may require continue:line %d\n",line);
-    stemRules.push_back(temp);
-  }
-  fclose(fp);
-  return 0;
+		*ch = 0;
+		temp.inflection = getInflectionNum(savech);
+		if ((ch = wcschr(savech = ch + 1, L';'))) *ch = 0;
+		temp.protect = wcsstr(savech, L"protect") != NULL;
+		temp.intact = wcsstr(savech, L"intact") != NULL;
+		temp.cont = wcsstr(savech, L"continue") != NULL;
+		temp.rulenum = line; /* Line number of rule in file */
+		/* Check replacement string for special 2-digit markers  */
+		int rl = temp.repstr.length();
+		if (rl > 1 && iswdigit(temp.repstr[rl - 1]) && iswdigit(temp.repstr[rl - 2]) && temp.cont == false)
+			lplog(L"** WARNING ** ReadRules: State marker may require continue:line %d\n", line);
+		stemRules.push_back(temp);
+	}
+	fclose(fp);
+	return 0;
 }
 
 int cStemmer::readPrefixRules(void)
-{ LFS
-  /* Format is: keystr,repstr where keystr and repstr are strings */
+{
+	LFS
+		/* Format is: keystr,repstr where keystr and repstr are strings */
 
-  FILE *fp=_wfopen(L"source\\lists\\prefixRules.txt",L"rb");
-  if (!fp) return NO_PREFIX_RULES_FILE;
-  wchar_t s[maxlinelength];
-  tPrefixRule temp;
-  int line;
-  /* Read a line at a time until eof */
-  for (line=1; fgetws(s,maxlinelength,fp); line++)
-  {
-		if (s[0]==0xFEFF) // detect BOM
-			memcpy(s,s+1,wcslen(s+1));
-    if ((s[0] == ';') || (s[0] == '\r') || (s[0] == '\n') || (s[0] == ' '))
-      continue;
-    wchar_t *ch=wcschr(s,L',');
-    if (!ch) 
+		FILE* fp = _wfopen(L"source\\lists\\prefixRules.txt", L"rb");
+	if (!fp) return NO_PREFIX_RULES_FILE;
+	wchar_t s[maxlinelength];
+	tPrefixRule temp;
+	int line;
+	/* Read a line at a time until eof */
+	for (line = 1; fgetws(s, maxlinelength, fp); line++)
+	{
+		if (s[0] == 0xFEFF) // detect BOM
+			memcpy(s, s + 1, wcslen(s + 1));
+		if ((s[0] == ';') || (s[0] == '\r') || (s[0] == '\n') || (s[0] == ' '))
+			continue;
+		wchar_t* ch = wcschr(s, L',');
+		if (!ch)
 		{
-		  fclose(fp);
+			fclose(fp);
 			return PREFIX_RULES_PARSE_ERROR;
 		}
-    *ch=0;
-    temp.keystr = s; /* Copy key string into the rule struct */
-    temp.repstr = ch+1;
-    temp.rulenum = line;
-    prefixRules.push_back(temp);
-  }
-  fclose(fp);
-  return 0;
+		*ch = 0;
+		temp.keystr = s; /* Copy key string into the rule struct */
+		temp.repstr = ch + 1;
+		temp.rulenum = line;
+		prefixRules.push_back(temp);
+	}
+	fclose(fp);
+	return 0;
 }
 
 bool sortRuleGreater(cStemmer::cSuffixRule a, cStemmer::cSuffixRule b)
-{ LFS
-  if (a.trail.count==b.trail.count)
-    return a.text.length() > b.text.length();
-  return a.trail.count < b.trail.count;
+{
+	LFS
+		if (a.trail.count == b.trail.count)
+			return a.text.length() > b.text.length();
+	return a.trail.count < b.trail.count;
 }
 
-int cStemmer::stem(MYSQL mysql, wstring word, vector <cSuffixRule> &rulesUsed,cIntArray &trail,int addRule)
-{ LFS
-  int state=s_continue;
-  int ret;
-  if (!stemRules.size() && (ret=readStemRules())<0) return ret;
-  if (addRule>=0) trail.push_back(addRule);
-  for (unsigned int r=0; r<stemRules.size(); r++)
-    if (state=applyStemRule(word,stemRules[r],rulesUsed,trail)==s_continue)
-      stem(mysql, rulesUsed[rulesUsed.size()-1].text,rulesUsed,trail,stemRules[r].rulenum);
-  if (addRule>=0) return rulesUsed.size();
-  if (ret=stripPrefix(mysql, word, rulesUsed)) 
+size_t cStemmer::stem(MYSQL mysql, wstring word, vector<cSuffixRule>& rulesUsed, cIntArray& trail, int addRule)
+{
+	LFS
+		int state = s_continue;
+	int ret;
+	if (!stemRules.size() && (ret = readStemRules()) < 0) return ret;
+	if (addRule >= 0) trail.push_back(addRule);
+	for (unsigned int r = 0; r < stemRules.size(); r++)
+		if (state = applyStemRule(word, stemRules[r], rulesUsed, trail) == s_continue)
+			stem(mysql, rulesUsed[rulesUsed.size() - 1].text, rulesUsed, trail, stemRules[r].rulenum);
+	if (addRule >= 0) return rulesUsed.size();
+	if (ret = stripPrefix(mysql, word, rulesUsed))
 		return ret;
-  sort(rulesUsed.begin(),rulesUsed.end(),sortRuleGreater);
-  return rulesUsed.size();
+	sort(rulesUsed.begin(), rulesUsed.end(), sortRuleGreater);
+	return rulesUsed.size();
 }
 
-bool cStemmer::isWordDBUnknown(MYSQL mysql,wstring word)
+bool cStemmer::isWordDBUnknown(MYSQL mysql, wstring word)
 {
 	tIWMM iWord = Words.query(word);
 	if (iWord != Words.end() && iWord->second.query(UNDEFINED_FORM_NUM) >= 0)
 		return true;
 	if (!myquery(&mysql, L"LOCK TABLES words w READ,wordForms wf READ")) return true;
 	wchar_t qt[QUERY_BUFFER_LEN_OVERFLOW];
-	_snwprintf(qt, QUERY_BUFFER_LEN, L"select COUNT(*) from words w,wordForms wf where w.id=wf.wordId and word=\"%s\" and wf.formId=%d", word.c_str(), UNDEFINED_FORM_NUM+1); // always add one when referring to DB formId
-	MYSQL_RES *result = NULL;
+	_snwprintf(qt, QUERY_BUFFER_LEN, L"select COUNT(*) from words w,wordForms wf where w.id=wf.wordId and word=\"%s\" and wf.formId=%d", word.c_str(), UNDEFINED_FORM_NUM + 1); // always add one when referring to DB formId
+	MYSQL_RES* result = NULL;
 	MYSQL_ROW sqlrow;
 	if (!myquery(&mysql, qt, result))
 	{
 		myquery(&mysql, L"UNLOCK TABLES");
 		return true;
 	}
-	int count=1;
+	int count = 1;
 	if ((sqlrow = mysql_fetch_row(result)) != NULL)
 		count = atoi(sqlrow[0]);
 	mysql_free_result(result);
@@ -249,47 +255,49 @@ bool cStemmer::isWordDBUnknown(MYSQL mysql,wstring word)
 	return count > 0;
 }
 
-int cStemmer::applyPrefixRule(MYSQL mysql, tPrefixRule r,vector <cSuffixRule> &rulesUsed,int originalSize,wstring word)
-{ LFS
-  // word must have sufficient length over the prefix as well as matching it over the prefix length.
-  if (word.length()<=r.keystr.length()+2 || r.keystr!=word.substr(0,r.keystr.length())) return 0;
-  cSuffixRule tmp;
-  tmp.text=word.substr(r.keystr.length(),word.length()-r.keystr.length());
+int cStemmer::applyPrefixRule(MYSQL mysql, tPrefixRule r, vector <cSuffixRule>& rulesUsed, int originalSize, wstring word)
+{
+	LFS
+		// word must have sufficient length over the prefix as well as matching it over the prefix length.
+		if (word.length() <= r.keystr.length() + 2 || r.keystr != word.substr(0, r.keystr.length())) return 0;
+	cSuffixRule tmp;
+	tmp.text = word.substr(r.keystr.length(), word.length() - r.keystr.length());
 	// does this word exist and is known?
-	if (isWordDBUnknown(mysql,tmp.text))
+	if (isWordDBUnknown(mysql, tmp.text))
 		return 0;
-  tmp.keystr=r.keystr;
-  tmp.repstr=r.repstr;
-  tmp.rulenum=-r.rulenum;
-  tmp.form=L"PREVIOUS";
-  rulesUsed.push_back(tmp);
-  for (int ru=0; ru<originalSize; ru++)
-  {
-    if (rulesUsed[ru].text.length()<=r.keystr.length()+2) continue; // won't leave any word left!
-    tmp.text=rulesUsed[ru].text.substr(r.keystr.length(),rulesUsed[ru].text.length()-r.keystr.length());
-    tmp.keystr=r.keystr;
-    tmp.repstr=r.repstr;
-    tmp.rulenum=-r.rulenum;
-    tmp.trail=rulesUsed[ru].trail;
-    tmp.trail.add(rulesUsed[ru].rulenum);
-    tmp.form=L"PREVIOUS";
-    rulesUsed.push_back(tmp);
-  }
-  return 0;
+	tmp.keystr = r.keystr;
+	tmp.repstr = r.repstr;
+	tmp.rulenum = -r.rulenum;
+	tmp.form = L"PREVIOUS";
+	rulesUsed.push_back(tmp);
+	for (int ru = 0; ru < originalSize; ru++)
+	{
+		if (rulesUsed[ru].text.length() <= r.keystr.length() + 2) continue; // won't leave any word left!
+		tmp.text = rulesUsed[ru].text.substr(r.keystr.length(), rulesUsed[ru].text.length() - r.keystr.length());
+		tmp.keystr = r.keystr;
+		tmp.repstr = r.repstr;
+		tmp.rulenum = -r.rulenum;
+		tmp.trail = rulesUsed[ru].trail;
+		tmp.trail.add(rulesUsed[ru].rulenum);
+		tmp.form = L"PREVIOUS";
+		rulesUsed.push_back(tmp);
+	}
+	return 0;
 }
 
 // prefixes are not nestable
 // apply all prefixes to all rulesUsed
-int cStemmer::stripPrefix(MYSQL mysql, wstring word, vector <cSuffixRule> &rulesUsed)
-{ LFS
-  if (!prefixRules.size() && readPrefixRules()<0) return -1;
-  int originalSize=rulesUsed.size();
-  for (unsigned int r=0; r<prefixRules.size(); r++)
-    applyPrefixRule(mysql,prefixRules[r],rulesUsed,originalSize,word);
-  return 0;
+int cStemmer::stripPrefix(MYSQL mysql, wstring word, vector <cSuffixRule>& rulesUsed)
+{
+	LFS
+		if (!prefixRules.size() && readPrefixRules() < 0) return -1;
+	size_t originalSize = rulesUsed.size();
+	for (unsigned int r = 0; r < prefixRules.size(); r++)
+		applyPrefixRule(mysql, prefixRules[r], rulesUsed, originalSize, word);
+	return 0;
 }
 
-int cStemmer::findLastFormInflection(vector <cSuffixRule> rulesUsed, vector <cSuffixRule>::iterator &r, wstring &form, int &inflection)
+int cStemmer::findLastFormInflection(vector <cSuffixRule> rulesUsed, vector <cSuffixRule>::iterator& r, wstring& form, int& inflection)
 {
 	LFS
 		form = L"ORIGINAL";
@@ -317,11 +325,11 @@ int cStemmer::findLastFormInflection(vector <cSuffixRule> rulesUsed, vector <cSu
 cStemmer::~cStemmer()
 {
 	LFS
-	stemRules.clear();
+		stemRules.clear();
 	prefixRules.clear();
 }
 
-bool cStemmer::wordIsNotUnknownAndOpen(tIWMM iWord,bool log)
+bool cStemmer::wordIsNotUnknownAndOpen(tIWMM iWord, bool log)
 {
 	if (unacceptableCombinationForms.empty())
 	{
@@ -390,7 +398,7 @@ bool cStemmer::wordIsNotUnknownAndOpen(tIWMM iWord,bool log)
 		if ((ucf = unacceptableCombinationForms.find(iWord->second.forms()[f])) != unacceptableCombinationForms.end())
 		{
 			if (log)
-			lplog(LOG_DICTIONARY, L"WordPosMAP %s is a %s.", iWord->first.c_str(), Forms[*ucf]->name.c_str());
+				lplog(LOG_DICTIONARY, L"WordPosMAP %s is a %s.", iWord->first.c_str(), Forms[*ucf]->name.c_str());
 			return false;
 		}
 	return (iWord->second.query(verbForm) >= 0) || (iWord->second.query(nounForm) >= 0) || (iWord->second.query(adverbForm) >= 0) || (iWord->second.query(adjectiveForm) >= 0);
